requirements.txt
asgiref==3.7.2
Django==4.2.7
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
requests==2.25.1
djangorestframework==3.14.0

gunicorn
